Refactor to eliminate old `ShieldClient`
