+++ broker/src/main/groovy/com/swisscom/cloud/sb/broker/backup/BackupOnShield.groovy
@@ -2,7 +2,7 @@ package com.swisscom.cloud.sb.broker.backup
 
 import com.swisscom.cloud.sb.broker.backup.shield.ShieldClient
 import com.swisscom.cloud.sb.broker.backup.shield.ShieldConfig
-import com.swisscom.cloud.sb.broker.backup.shield.ShieldServiceConfig
+import com.swisscom.cloud.sb.broker.backup.shield.BackupParameter
 import com.swisscom.cloud.sb.broker.backup.shield.ShieldTarget
 import com.swisscom.cloud.sb.broker.model.Parameter
 import com.swisscom.cloud.sb.broker.model.ServiceInstance
@@ -12,7 +12,7 @@ import org.springframework.beans.factory.annotation.Autowired
 
 @CompileStatic
 trait BackupOnShield {
-    private final String SERVICE_DEFINITION_PARAMETER_BACKUP_PREFIX = "BACKUP_"
+    private final String PLAN_PARAMETER_BACKUP_PREFIX = "BACKUP_"
 
     @Autowired
     ProvisioningPersistenceService provisioningPersistenceService
@@ -27,13 +27,21 @@ trait BackupOnShield {
 
     abstract String shieldAgentUrl(ServiceInstance serviceInstance)
 
-    ShieldServiceConfig shieldServiceConfig(ServiceInstance serviceInstance) {
-        shieldServiceConfigFromParameters(serviceInstance.plan.parameters)
+    def backupJobName(String jobPrefix, String serviceInstanceId) {
+        "${jobPrefix}${serviceInstanceId}"
     }
 
-    ShieldServiceConfig shieldServiceConfigFromParameters(Set<Parameter> parameters) {
+    def backupTargetName(String targetPrefix, String serviceInstanceId) {
+        "${targetPrefix}${serviceInstanceId}"
+    }
+
+    BackupParameter getBackupParameter(ServiceInstance serviceInstance) {
+        getBackupParameter(serviceInstance.plan.parameters)
+    }
+
+    BackupParameter getBackupParameter(Set<Parameter> parameters) {
         def planParamtersForBackup = parameters.findAll {
-            it.getName().startsWith(SERVICE_DEFINITION_PARAMETER_BACKUP_PREFIX)
+            it.getName().startsWith(PLAN_PARAMETER_BACKUP_PREFIX)
         }
         String scheduleName = planParamtersForBackup.find {
             it.getName().equals("BACKUP_SCHEDULE_NAME")
@@ -44,6 +52,6 @@ trait BackupOnShield {
         String storageName = planParamtersForBackup.find {
             it.getName().equals("BACKUP_STORAGE_NAME")
         }?.getValue()
-        new ShieldServiceConfig(scheduleName: scheduleName, retentionName: policyName, storeName: storageName)
+        new BackupParameter(scheduleName: scheduleName, retentionName: policyName, storeName: storageName)
     }
 }
