+++ broker/src/main/groovy/com/swisscom/cloud/sb/broker/backup/shield/ShieldBackupRestoreProvider.groovy
@@ -1,6 +1,6 @@
 package com.swisscom.cloud.sb.broker.backup.shield
 
-import com.swisscom.cloud.sb.broker.backup.UserBackupOnShield
+import com.swisscom.cloud.sb.broker.backup.BackupRestoreProvider
 
-trait ShieldBackupRestoreProvider extends UserBackupOnShield {
+trait ShieldBackupRestoreProvider extends BackupRestoreProvider {
 }
